move schemas under specs
