+++ src/aioway/datasets/frames/chunks.py
@@ -10,7 +10,7 @@ import numpy as np
 import tensordict as td
 
 from aioway._utils import is_list_of
-from aioway.schemas import AttrSet, attr_set
+from aioway.specs import AttrSet, attr_set
 
 from .frames import Frame
