+++ src/aioway/io/audios.py
@@ -10,7 +10,7 @@ from torch.utils import data
 from torchcodec import decoders as dec
 
 from aioway._torch import current_fake_mode, torch_set_fake_mode_func
-from aioway.schemas import Attr
+from aioway.specs import Attr
 from aioway.tags import SampleRateTag
 from aioway.tags.media import IsStftTag
