+++ src/aioway/io/images.py
@@ -11,7 +11,7 @@ from torchvision import io as vio
 from torchvision.transforms import v2 as tt
 
 from aioway._torch import current_fake_mode, torch_set_fake_mode_func
-from aioway.schemas import AttrLike, attr
+from aioway.specs import AttrLike, attr
 from aioway.tags import IsImageTag
 
 from ._bases import TorchCompatible
