+++ src/aioway/io/videos.py
@@ -10,7 +10,7 @@ from torchcodec import decoders as dec
 
 from aioway._torch import current_fake_mode, torch_set_fake_mode_func
 from aioway._utils import num_threads
-from aioway.schemas import Attr
+from aioway.specs import Attr
 from aioway.tags import IsVideoTag
 
 from ._av import VideoStream
