+++ src/aioway/specs/__init__.py
@@ -1,0 +1,3 @@
+# Copyright (c) AIoWay Authors - All Rights Reserved
+
+from .schemas import *
