+++ src/aioway/specs/schemas/_bases.py
