+++ src/aioway/specs/schemas/attrs.py
