+++ src/aioway/specs/schemas/devices.py
