+++ src/aioway/specs/schemas/shapes.py
